Separated audio player from main page
