+++ src/routes/our-music/audioData.js
@@ -20,4 +20,26 @@ export const audioData = [
         name:
             'How to Mug Week 3 Especially During Tutorials in NUS by Nicholas Russell Saerang',
     },
+    {
+        link: 'https://sveltejs.github.io/assets/music/strauss.mp3',
+        name: 'Literally The Blue Danube Waltz',
+    },
+    {
+        link: 'https://sveltejs.github.io/assets/music/satie.mp3',
+        name: 'Literally Gymnopédie no. 1',
+    },
+    {
+        link: 'https://sveltejs.github.io/assets/music/beethoven.mp3',
+        name: 'Literally Symphony no. 5 in Cm, Op. 67 - I. Allegro con brio',
+    },
+    {
+        link: 'https://sveltejs.github.io/assets/music/mozart.mp3',
+        name:
+            'Literally Requiem in D minor, K. 626 - III. Sequence - Lacrymosa',
+    },
+    {
+        link: 'https://sveltejs.github.io/assets/music/holst.mp3',
+        name:
+            'The Best End Credit Music : Mars, the Bringer of War by Gustav Holst',
+    },
 ];
